daemon: Set the volume based on property, volume value or volume key.
Every event defined in the configuration has it's own volume
control in the form of "x-maemo-ngf-<event>". This can be overridden
with property volume_role.

The value for the volume role is set before sound is played and value
comes from override, volume_set or the value from the profiles.