event: Audio repeating indefinitely or by a number of times.
User can specify "tone_repeat" boolean property the enable repeating of
the audio indefinitely. This will cause the event not complete until it
is stopped or the max_length of the event is reached.

"tone_repeat_count" integer property allows us to specify the number of
times the tone is repeated. Once the repeat count is reached, audio part
is considered as complete.