immvibe: fix custom vibration pattern loading
There are three special cases in addition to the normal vibration
loading. In these cases, we need to figure out the sound for which
the vibration pattern will be loaded.

1. custom sound provided by user
2. sound filename from the active event
3. sound filename from a specified context key

After the sound has been resolved, we can resolve if the sound
filename belongs to "factory-shipped" sounds, ie. it has a custom
vibration pattern. Then we try to load it.

If these fail, then we'll use the default vibration defined as
a filename or profile key in event configuration.