ffmemless: Add effect repeat support
Added effect repeat support to ffmemless plugin. The plugin detects if
sound.repeat is TRUE and plays the effect INT32MAX times (effectively
infinitely for user). There is also new settings parameter
"_REPEAT" to create single shot effects that consist of multiple
vibrations.

This patch fixes also a issue with the plugin not reporting
the end of playback on single shot effects.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jollamobile.com>