ffmemless: Rename settings environment variable
The original plan for the settings environment variable was that
we could use the same settings file for qtfeedback and ngfd.
Unfortunately QSettings cannot read glib's .ini file format,
and vice versa glib cannot comprehend settings values without
bracket header.

So, change the ngf environment variable to be different from
the qtfeedback one, so that separate device/distro specific values
can be set in separate files for qtfeedback and ngfd.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jollamobile.com>