[ffmemless] Allow effects to be defined as touch screen related
Added new configuration option for effects that allow them to be
classified as touch screen related. For touch screen related effects
the touchscreen.vibrate.level is followed to determine if the effect
should be played or not.

The option can be enabled in ffmemless configuration file by adding
a "_TOUCH" postfixed configuration line. E.g. one could define NGF_SHORT
to be a touch screen related effect by adding to the .ini file:

NGF_SHORT_TOUCH = 1

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jollamobile.com>